# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AquaForge API",
    "description": "Swim Meet Optimization API - Backend services for AquaForge.ai",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analytics/compare": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Compare Teams",
        "description": "Compare two teams across all events.\n\nReturns head-to-head analysis and event-by-event breakdown.",
        "operationId": "compare_teams_api_v1_analytics_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_compare_teams_api_v1_analytics_compare_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/swimmer/{swimmer_name}": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Analyze Swimmer",
        "description": "Get detailed analytics for a specific swimmer.",
        "operationId": "analyze_swimmer_api_v1_analytics_swimmer__swimmer_name__post",
        "parameters": [
          {
            "name": "swimmer_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swimmer Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                },
                "title": "Team Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/event/{event_name}": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Analyze Event",
        "description": "Get detailed analytics for a specific event.",
        "operationId": "analyze_event_api_v1_analytics_event__event_name__post",
        "parameters": [
          {
            "name": "event_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_event_api_v1_analytics_event__event_name__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/depth": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Analyze Depth",
        "description": "Analyze team depth across all events.",
        "operationId": "analyze_depth_api_v1_analytics_depth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array",
                "title": "Team Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_analyze_event_api_v1_analytics_event__event_name__post": {
        "properties": {
          "seton_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Seton Data"
          },
          "opponent_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Opponent Data"
          }
        },
        "type": "object",
        "required": [
          "seton_data",
          "opponent_data"
        ],
        "title": "Body_analyze_event_api_v1_analytics_event__event_name__post"
      },
      "Body_compare_teams_api_v1_analytics_compare_post": {
        "properties": {
          "seton_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Seton Data"
          },
          "opponent_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Opponent Data"
          }
        },
        "type": "object",
        "required": [
          "seton_data",
          "opponent_data"
        ],
        "title": "Body_compare_teams_api_v1_analytics_compare_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}